optimize(os): Optimize the directory and code structure of the os section.
