+++ packages/client/config/main.ts
@@ -1,56 +1,18 @@
-import { app, BrowserWindow, Tray } from "electron";
-import path from "path";
+import { BrowserWindow, app } from "electron";
 import "../../automation/index";
-import { init } from "../module/os/windows";
+import os from "../module/os";
 
 // Handle creating/removing shortcuts on Windows when installing/uninstalling.
 if (require("electron-squirrel-startup")) {
   app.quit();
 }
-
-init();
-
-const createWindow = () => {
-  // Create the browser window.
-  const mainWindow = new BrowserWindow({
-    icon: "assets/icon.png",
-    title: "Any Design - 原型直接生成App",
-    width: 1440,
-    height: 900,
-    titleBarStyle: "hidden",
-    darkTheme: true,
-    autoHideMenuBar: true,
-    show: false,
-    resizable: false,
-    webPreferences: {
-      preload: path.join(__dirname, "preload.js"),
-      // contextIsolation: false,
-      // nodeIntegration: true,
-    },
-  });
-
-  // and load the index.html of the app.
-  if (MAIN_WINDOW_VITE_DEV_SERVER_URL) {
-    mainWindow.loadURL(MAIN_WINDOW_VITE_DEV_SERVER_URL);
-  } else {
-    mainWindow.loadFile(
-      path.join(__dirname, `../renderer/${MAIN_WINDOW_VITE_NAME}/index.html`)
-    );
-  }
-
-  // Jump white screen the application init.
-  mainWindow.webContents.addListener("did-finish-load", () => {
-    mainWindow.show();
-  });
-
-  // Open the DevTools.
-  // mainWindow.webContents.openDevTools();
-};
+// 注册沙盒事件
+os.window.registerWindowEvent();
 
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-app.on("ready", createWindow);
+app.on("ready", os.window.createMainWindow);
 
 // Quit when all windows are closed, except on macOS. There, it's common
 // for applications and their menu bar to stay active until the user quits
@@ -65,7 +27,7 @@ app.on("activate", () => {
   // On OS X it's common to re-create a window in the app when the
   // dock icon is clicked and there are no other windows open.
   if (BrowserWindow.getAllWindows().length === 0) {
-    createWindow();
+    os.window.createMainWindow();
   }
 });
