+++ packages/client/module/os/index.ts
@@ -1,0 +1,5 @@
+import * as window from "./windows";
+
+export default {
+  window: window,
+};
