+++ packages/client/module/os/windows/index.ts
@@ -1,0 +1,11 @@
+import {
+  registerCreateWindowEvent,
+  registerOpenModalEvent,
+} from "./events/window-events";
+
+export * from "./main/main-window";
+
+export function registerWindowEvent() {
+  registerOpenModalEvent();
+  registerCreateWindowEvent();
+}
